[Issue #792] Removes compilation warnings from libs visionlib/imgAnalyze

+++ src/libs/visionlib/imgAnalyze/linesDetection.cpp
@@ -93,7 +93,7 @@ namespace visionLibrary {
 			i = 0;
 			first = true;
 
-			while ((i < (vecContours[contour].size() - (i_jump - 1)))) {
+			while ((static_cast<unsigned int>(i) < (vecContours[contour].size() - (i_jump - 1)))) {
 
 				counter++;
